Move `routes` to Console
