+++ features/admin.core.v1/configs/index.ts
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020, WSO2 LLC. (https://www.wso2.com). All Rights Reserved.
+ * Copyright (c) 2020-2024, WSO2 LLC. (https://www.wso2.com).
  *
  * WSO2 LLC. licenses this file to you under the Apache License,
  * Version 2.0 (the "License"); you may not use this file except
@@ -23,5 +23,4 @@
  */
 export * from "./app";
 export * from "./documentation";
-export * from "./routes";
 export * from "./ui";
